backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pandas
scikit-learn
numpy
python-multipart
# For data generation if needed
faker
certifi
